Made refactoring
